Fix announcements not saving; redesign admin panel
Announcements silently failed to sync to Supabase because sort_order
used Date.now() (a 13-digit ms timestamp) against an `integer` column,
overflowing it — switched the column to bigint.

Also replaced the pill-shaped admin bar (which turned into a messy
blob once it wrapped to multiple rows) with a proper collapsible card:
a header with sync status, a primary edit action, a 2-column grid for
secondary actions, and dangerous/exit actions visually separated.

+++ supabase-schema.sql
@@ -24,7 +24,7 @@ create table if not exists announcements (
   body text not null default '',
   date_label text not null default '',
   image_url text,
-  sort_order integer not null default 0,
+  sort_order bigint not null default 0,
   created_at timestamptz not null default now()
 );
